v1.85.10

+++ enkf/common/observations.h
@@ -43,7 +43,7 @@ typedef struct {
     short int product;
     short int instrument;
     short int fid;
-    short int batch;
+    int batch;
     double value;
     double std;
     double lon;
